github: Add cryptography dependencies
Not all pypy releases have a cryptography wheel prebuilt, especially the
newer ones.

In such a case, cryptography would need to be built and it comes with a
bunch of dependencies for the build to succeed, in particular Rust
support.

Let's make sure we have the proper build dependencies when using pypy.

Signed-off-by: Maxime Ripard <mripard@kernel.org>